# Optimize TryParse out bool extension method by splitting slow path