This type of optimization can be applied in a lot of other APIs.

**~5-10% improvement for the single segment case.**

``` ini

BenchmarkDotNet=v0.11.1, OS=Windows 10.0.17763
Intel Core i7-6700 CPU 3.40GHz (Skylake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=3.0.100-alpha1-009667
  [Host]     : .NET Core 2.1.3 (CoreCLR 4.6.26725.06, CoreFX 4.6.26725.05), 64bit RyuJIT
  DefaultJob : .NET Core 2.1.3 (CoreCLR 4.6.26725.06, CoreFX 4.6.26725.05), 64bit RyuJIT


```
|                                       Method |     Mean |     Error |    StdDev | Scaled | Allocated |
|--------------------------------------------- |---------:|----------:|----------:|-------:|----------:|
|           SingleSegmentBufferReader_original | 5.963 us | 0.0477 us | 0.0446 us |   1.00 |       0 B |
| SingleSegmentBufferReader_original_with_opts | 5.376 us | 0.0449 us | 0.0420 us |   0.90 |       0 B |