@ahsonkhan 2018-10-23
Master got better (because of https://github.com/dotnet/corefxlab/pull/2551):

|                             Method |     Mean |     Error |    StdDev | Scaled | Allocated |
|----------------------------------- |---------:|----------:|----------:|-------:|----------:|
| SingleSegmentBufferReader_original | 5.637 us | 0.0911 us | 0.0807 us |   1.00 |       0 B |
|      SingleSegmentBufferReader_new | 5.428 us | 0.0369 us | 0.0308 us |   0.96 |       0 B |