@ilmax 2018-10-23
@ahsonkhan out of curiosity, does a go-to is faster than two returns?