@ahsonkhan 2018-10-23
> out of curiosity, does a go-to is faster than two returns?

No, not really (at least not visible in a micro-benchmark). However, the disassembly is smaller: https://www.diffchecker.com/93DU926Y