@JeremyKuhne 2018-10-23
@AndyAyersMS do you have any comments on this difference? How valuable is the reduction in instruction count here? Is this something we could theoretically get to with JIT improvements and no goto?

> No, not really (at least not visible in a micro-benchmark). However, the disassembly is smaller: https://www.diffchecker.com/93DU926Y