@AndyAyersMS 2018-10-23
Unless this is a generic method/class with a large number of potential instantiations (does not seem to be the case here), I would not worry about saving few bytes of code and would favor whatever is more readable or natural looking.

The jit behavior for returns is dependent on target architecture and you may well see different impacts for x64 or arm64. Likely at some point we'll adjust the jit strategy around multiple returns (see for instance dotnet/coreclr#13814) as we know there are some quirks in the current jit approach. Once we tackle that, it generally won't matter how you express this in your code.