@JeremyKuhne 2018-10-23
Thanks @AndyAyersMS. @ahsonkhan, given these comments I'd not put `goto` in unless we measure an important difference in an upstream scenario or see it in something really core like `Advance`.